Add files via upload
This project is made using Core Java and GUI (Swings). In this project, the customer first logs in via his/her account and then choose the source(to board the train or flight) and destination where he finally lands. All this is done with the help of the drop-down menu in which the number of cities of source and destination is fixed. Accordingly, the price is also calculated. For, example if a person goes from Delhi to Jaipur he has to pay less, but if he goes from Delhi to Chennai he has to pay more amount.  If by chance customer wants to go from Delhi to Delhi the price shown is Rs.0 on the screen. Same is the case if a person wants to travel via train. After this, a window appears in which the customer enters his details of the virtual credit card- credit card number and the CVV number. The amount shown to the customer will be deducted from credit card and the message will be shown on the window. All the files to the code have been attached.